Insert Product's Image

+++ database/migrations/2024_07_23_154647_create_orders_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->double('total', 8, 2);
+            $table->unsignedBigInteger('total');
             $table->timestamps();
 
             $table->foreignId('customer_id')->constrained('customers');
